fix Bottled Memory and some refactoring. Refactoring probably should have been a separate commit, but alas, I did not do that.
